+++ ECom-Classes/src/main/java/com/ad/ecom/orders/repository/OrderItemRepository.java
@@ -1,6 +1,6 @@
 package com.ad.ecom.orders.repository;
 
-import com.ad.ecom.orders.persistence.OrderItem;
+import com.ad.ecom.orders.persistance.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
