+++ ECom-Classes/src/main/java/com/ad/ecom/orders/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.ad.ecom.orders.repository;
 
-import com.ad.ecom.orders.persistence.Order;
+import com.ad.ecom.orders.persistance.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
