+++ ECom-Classes/src/main/java/com/ad/ecom/products/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.ad.ecom.products.repository;
 
-import com.ad.ecom.products.persistence.Product;
+import com.ad.ecom.products.persistance.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
