+++ ECom-Classes/src/main/java/com/ad/ecom/user/cart/repository/CartItemRepository.java
@@ -1,6 +1,6 @@
 package com.ad.ecom.user.cart.repository;
 
-import com.ad.ecom.user.cart.persistence.CartItem;
+import com.ad.ecom.user.cart.persistance.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
