+++ ECom-Core/src/main/java/com/ad/ecom/core/registration/service/impl/RegistrationServiceImpl.java
@@ -1,15 +1,15 @@
 package com.ad.ecom.core.registration.service.impl;
 
-import com.ad.ecom.common.ResponseMessage;
+import com.ad.ecom.common.dto.ResponseMessage;
 import com.ad.ecom.common.stub.ResponseType;
 import com.ad.ecom.core.registration.service.RegistrationService;
 import com.ad.ecom.core.util.WebTemplates;
 import com.ad.ecom.core.registration.util.emailEvent.VerificationEmailEvent;
-import com.ad.ecom.ecomuser.persistence.EcomUser;
+import com.ad.ecom.ecomuser.persistance.EcomUser;
 import com.ad.ecom.ecomuser.repository.EcomUserRepository;
 import com.ad.ecom.exception.UserAlreadyExistsException;
 import com.ad.ecom.registration.dto.RegistrationRequest;
-import com.ad.ecom.registration.persistence.VerificationToken;
+import com.ad.ecom.registration.persistance.VerificationToken;
 import com.ad.ecom.registration.repository.VerificationTokenRepository;
 import com.ad.ecom.util.EComUtil;
 import org.springframework.beans.factory.annotation.Autowired;
