+++ ECom-Core/src/main/java/com/ad/ecom/core/registration/util/emailEvent/VerificationEmailEvent.java
@@ -1,6 +1,6 @@
 package com.ad.ecom.core.registration.util.emailEvent;
 
-import com.ad.ecom.ecomuser.persistence.EcomUser;
+import com.ad.ecom.ecomuser.persistance.EcomUser;
 import lombok.Getter;
 import org.springframework.context.ApplicationEvent;
