+++ ECom-Orders/src/main/java/com/ad/ecom/service/impl/OrdersServiceImpl.java
@@ -1,29 +1,29 @@
 package com.ad.ecom.service.impl;
 
-import com.ad.ecom.common.ResponseMessage;
-import com.ad.ecom.common.dto.Item;
+import com.ad.ecom.common.dto.ResponseMessage;
+import com.ad.ecom.products.dto.Item;
 import com.ad.ecom.common.stub.ResponseType;
 import com.ad.ecom.core.context.EComUserLoginContext;
-import com.ad.ecom.discounts.persistence.DiscountSubscription;
+import com.ad.ecom.discounts.persistance.DiscountSubscription;
 import com.ad.ecom.discounts.repository.DiscountSubscriptionRepository;
-import com.ad.ecom.ecomuser.persistence.EcomUser;
+import com.ad.ecom.ecomuser.persistance.EcomUser;
 import com.ad.ecom.ecomuser.repository.EcomUserRepository;
 import com.ad.ecom.ecomuser.stubs.Role;
 import com.ad.ecom.orders.dto.OrderInfo;
 import com.ad.ecom.orders.dto.OrderTrackingInfo;
-import com.ad.ecom.orders.persistence.Order;
-import com.ad.ecom.orders.persistence.OrderItem;
+import com.ad.ecom.orders.persistance.Order;
+import com.ad.ecom.orders.persistance.OrderItem;
 import com.ad.ecom.orders.repository.OrderRepository;
 import com.ad.ecom.orders.stubs.OrderEvent;
 import com.ad.ecom.orders.stubs.OrderStatus;
-import com.ad.ecom.products.persistence.Product;
+import com.ad.ecom.products.persistance.Product;
 import com.ad.ecom.products.repository.ProductRepository;
 import com.ad.ecom.service.OrdersService;
 import com.ad.ecom.ssm.OrdersSMInterceptor;
-import com.ad.ecom.user.cart.persistence.Cart;
-import com.ad.ecom.user.cart.persistence.CartItem;
+import com.ad.ecom.user.cart.persistance.Cart;
+import com.ad.ecom.user.cart.persistance.CartItem;
 import com.ad.ecom.user.cart.repository.CartRepository;
-import com.ad.ecom.user.profile.persistence.Address;
+import com.ad.ecom.user.profile.persistance.Address;
 import com.ad.ecom.user.profile.repository.AddressRepository;
 import com.ad.ecom.util.DateConverter;
 import com.ad.ecom.util.SSMUtil;
