+++ ECom-Products/src/main/java/com/ad/ecom/discounts/service/impl/DiscountsServiceImpl.java
@@ -1,12 +1,12 @@
 package com.ad.ecom.discounts.service.impl;
 
-import com.ad.ecom.common.ResponseMessage;
+import com.ad.ecom.common.dto.ResponseMessage;
 import com.ad.ecom.common.stub.ResponseType;
 import com.ad.ecom.discounts.dto.DiscountsDto;
 import com.ad.ecom.discounts.dto.DiscountsFilter;
 import com.ad.ecom.discounts.dto.DiscountsObjDto;
-import com.ad.ecom.discounts.persistence.Discount;
-import com.ad.ecom.discounts.persistence.QDiscount;
+import com.ad.ecom.discounts.persistance.Discount;
+import com.ad.ecom.discounts.persistance.QDiscount;
 import com.ad.ecom.discounts.repository.DiscountRepository;
 import com.ad.ecom.discounts.service.DiscountsService;
 import com.ad.ecom.discounts.stubs.DiscountStatus;
