+++ ECom-User/src/main/java/com/ad/ecom/user/profile/util/emailEvent/ProfileEmailEventsListener.java
@@ -1,7 +1,7 @@
 package com.ad.ecom.user.profile.util.emailEvent;
 
-import com.ad.ecom.ecomuser.persistence.EcomUser;
-import com.ad.ecom.registration.persistence.VerificationToken;
+import com.ad.ecom.ecomuser.persistance.EcomUser;
+import com.ad.ecom.registration.persistance.VerificationToken;
 import com.ad.ecom.registration.repository.VerificationTokenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.event.EventListener;
